Fragmenta a varios ficheros el algoritmo de inicialización #116 Ahora según la cabecera se encuentra en una carpeta u otra
